Redo the XmppConnection API in a more push/pull manner
Start of the refactored API for xmpp connection. The intention is that the
XmppConnection is a very thin layer over GIOStream and the XMPP reader/writers.
The actual scheduling of stanzas and a more advanced API for adding filters etc
will be part of a higher level object.

This patch has stubs for all the new methods with the old ones commented out so
tests can compile and run but fail.