Fixes calls to wocky_xmpp_connection_force_close_async. DEBUG gives file+line.
Never put more than 1 wocky_xmpp_connection_force_close_async in flight for a
given connection.

g_object_ref the wocky porter passed in as the user_data to the call so that
it is guaranteed to still be alive when the callback happens.

Report an error in idle if we try to force a shutdown twice.

DEBUG gives file + line now.