Correctly ship test certificates in out-of-tree builds
Previously the wildcards were matching no files, because they were being
applied to the builddir (where the Makefile is) rather than the srcdir.

I tried to do someting fancy involving $(CERTS), but $(CERTS) contains
absolute paths (for reasonably good reasons: they're passed as #defines
into a test binary) so then I tried mangling them back into relative
paths in $(EXTRA_DIST) but then I decided this was clearer.

Including directories in $(EXTRA_DIST) is discouraged by the Automake
documentation, and is unnecessary in this case because the contents of
certs/cas and certs/crl that we want are in $(EXTRA_DIST) anyway.