wocky/Makefile.am: fix generation of *-enumtypes.h
Previously, if foo-enumtypes.h was built as a result of a dependency from
another directory, we'd try to use the full path
(/home/smcv/.../wocky/foo) in the macro names, which is both syntactically
invalid and not what we want. The GNU make $(notdir) operator fixes this.

Also, the previous rule was somewhat illegible; factoring out the string
manipulation into temporary variables improves this, I think.