Make multi-line variables consistent
The right-aligned backslashes are great and all, but when you add a new
header that's longer (as I did with wocky-pubsub-service-protected.h)
you have to adjust the whole list or it becomes inconsistent.

Also the wocky-muc.c entry had a whole bunch of tabs rather than spaces
which made it not line up anyway.