Add a configure switch to install headers
Previously, Gabble hard-coded a list of Wocky headers to install if it
was built with --enable-plugin-api. This is fragile in the face of the
list of headers changing. With this patch, passing
--with-installed-headers=/path/to/headers to Wocky's configure will
cause `make install` to put headers there.

I had to split BUILT_SOURCES and HANDWRITTEN_SOURCES into sources and
headers, and lowercase them because FOO_HEADERS is magic in Automake. I
took the liberty of sorting the lists while I was at it.

This patch originally didn't work, and was made to work by Mike Ruprecht
<mike.ruprecht@collabora.co.uk>, but his fixes languished waiting for me
to review them for long enough that wocky/Makefile.am had changed so
much that the patch needed rewriting. Sorry. :(