Deal with empty replies to create_node()
When you ask a pubsub server to create you a node, in some circumstances
the IQ result contains <pubsub><create node='$NodeID'/></>, and in some
circumstances it does not:

• If you included a NodeID in your request, and the server did exactly
  as you asked and created that node, there's no need to include any
  information in the result's body as you already know what it would
  contain. (XEP-0060 actually goes further than this, and says that the
  result MUST NOT have a body in this case, which seems overzealous.)

• If you included a NodeID in your request, but the server created a
  node with a different NodeID, then it MUST include a body in the
  result to tell you the node's actual ID.

• If you did not include a NodeID in your request (to create an "instant
  node"), the server MUST include a body in the result to tell you the
  ID it decided to use.

We previously assumed that the result contains a body, so the first case
above wouldn't work with a fully-pedantic server. This patch fixes the
tests to test that case, fixes the implementation to support that case,
and adds a test for the second case.