Allow fields' type to be unspecified when submitting.
When we parse incoming forms, we never produce fields with unspecified
type: if the XML doesn't specify the type, we default to text-single as
specified by XEP-0004. This is not completely satisfactory: when parsing
a form result, we should be able to pre-seed the WockyDataForm object
with prior knowledge about the types of fields. But it'll do for now.

When submitting a form, we allow fields to have unspecified type, and
submit them with no type='' attribute. Quoth XEP-0004:

> For data forms of type "submit", inclusion of the 'type' attribute is
> OPTIONAL, since the form-processing entity is assumed to understand
> the data types associated with forms that it processes.