Add helpers to set data form fields' values
Rootling around in a hash table to fish out the right struct and then
replacing one of its members looked kind of strange. This is more
obvious and easier. The API also supports filling in fields that the
WockyDataForm object doesn't know about; this is to support blindly
submitting forms with well-known fields (such as PubSub configuration)
without having to first ask the server for the form.

This patch also adds a test for such blind submission.