Consistently use 'self' in DataForm, PubsubNode, PubsubService
These classes sometimes used 'self' as the first parameter to object
methods, and sometimes used 'form', 'node', 'service', etc. In some
cases, the parameter name was different between the declaration and
definition. This patch makes them always use 'self', and tweaks the
gtkdoc comments to match (some of which matched the definition rather
than the declaration, which makes gtkdoc cry).