Use G_GNUC_NULL_TERMINATED for build spec-cy functions
Unfortunately, the functions' previous signatures were awkward for
NULL-termination: because the first tag in the specification was an
explicit non-pointer argument, if you didn't actually want the stanza to
have a body you had to say:

  wocky_xmpp_stanza_build_iq_result (..., WOCKY_STANZA_END, NULL);

You couldn't use NULL in place of WOCKY_STANZA_END because WockyBuildTag
isn't a pointer type; and you couldn't omit it because then the varargs
aren't NULL-terminated. So, this patch removes the explicit first tag
argument. While this may be slightly worse from a documentation point of
view, I think it's better than having to document the special case
above.