Deconstruct the PubSub distillery a bit.
This exposes a function which accepts a stanza that you've already
verified is an iq type=result, and pulls out a child node. Previously,
if you weren't happy with wocky_pubsub_distill_iq_reply() and friends'
conversion from an <error/> element to a GError, you had to re-do the
breakdown, because these guts weren't exposed.