Return node trees from the distillery
Previously all these functions were technically incorrect: they were
calling wocky_porter_send_iq_finish(), and if it yielded a stanza,
they'd grab a WockyNode * out of the stanza, *unref the stanza*, and
then return the pointer to the caller. So, this relied on the stanza
happening to stay alive long enough for the caller to frob the nodes.

This makes some code a bit more verbose, but not much.