connector: assert that connect_to_service_finish() isn't broken.
Coverity was upset because we dereferenced 'error' just before checking
if it's NULL. I think it's right to be upset. But if
g_socket_client_connect_to_service_finish() returned NULL, either it set
'error', or it's buggy, or we're buggy and called it inappropriately. So
let's blow up if it's NULL in this case.