Remove certificate re-creation rules, they tend to break
The Makefile had some voodoo to re-create the various certificates used for
testing if needed.. Unfortunately they tend to break. Also these rules will
only be needed in 2037. If we're still using SSL and Wocky at that point, then
figuring out how to update the certificates to be valid a little bit longer
shouldn't be that much work :)