Pubsub: unleak calls to g_simple_async_result_get_source_object()
I tracked down a bug in Gabble to not unreffing the result of
g_simple_async_result_get_source_object() and thought I'd check Wocky.
I'm proud to report that I am responsible for all instances of this
error in Wocky. I'm even prouder to report that I found it once before,
added a comment explaining it, and then promptly forgot about it IN THE
SAME FILE. Dunce hat for me.

Reviewed-by: Simon McVittie <simon.mcvittie@collabora.co.uk>