Work around chat.facebook.com's broken SRV record.
_xmpp-client._tcp.chat.facebook.com returns a CNAME record, not a SRV
record. (But only on some connections! Apparently some people just get
nothing back, or something.) Anyway, if you get unlucky, this mean Wocky
just gave up and threw an error.

Instead, if a domain has an SRV record and we fail to connect to it, we
now fall back to connecting to the domain directly; if the latter fails
too, we report the original error. If a domain has no SRV record, we
fall back fully, and the behaviour is unchanged.

Fixes: <https://bugs.freedesktop.org/show_bug.cgi?id=28051>
Reviewed-by: Simon McVittie <simon.mcvittie@collabora.co.uk>