Fix tests for chat.facebook.com workaround.
The test case matching the chat.facebook.com situation was expected to
fail to connect, but the workaround made connecting succeed, breaking
the test suite. I'm a bad person.

So. I repurposed that test case to be one which is supposed to succeed,
and made a clone for a slightly different situation which should still
fail to connect.