Don't retry indefinitely on failed OpenSSL writes
Previously, the response to any partial write was to try again later, even if
the error returned indicated that it would never succeed. Instead, we should
only retry later if WOULD_BLOCK is returned (indicating only that the write
buffer is full, as opposed to an actual error).

This error was only in the OpenSSL backend; gnutls does not require us to
implement this logic at all.