Add a GSource wrapping libiphb.
The Loudmouth patch to add libiphb support for Maemo 5 intertwingled it
with other keepalive logic, but I think separating it out makes it
cleaner.

libiphb is part of DSME (http://gitorious.org/dsme/) which is now LGPL
v2.1 (note: not "or later"). This is an optional dependency, so doesn't
impact Wocky's license in general.

Currently the source is completely inert if we don't have iphb support;
it'll be fixed to fall back to acting as a timeout source later.