Replace WockyPing's timeouts with WockyHeartbeat
This strictly speaking introduces a regression: we no longer delay the
next ping whenever we send data. But actually, I don't think this is
all that necessary: if we want to do something in response to getting no
replies to our pings, we have to be sure we're sending them in the first
place. They're only tiny.