Test sending a stanza just before destroying porter.
I'm astonished that this didn't have a test already.

If we call wocky_porter_send_async(), then destroy the porter, and only
then spin the main loop, we'd crash. This is because after the porter
died, it'd take the xmpp connection down, which would then fire a
callback to tell the porter that sending a stanza failed. The fix is
simple: ref the porter across calls to
wocky_xmpp_connection_send_stanza_async().

Reviewed-by: Nicolas Dufresne <nicolas.dufresne@collabora.co.uk>
Fixes: <https://bugs.freedesktop.org/show_bug.cgi?id=30986>