CapsCache: don't potentially double-free statement
caps_cache_bind_*() call sqlite3_finalize() on the statement you pass in
if binding fails. So we should not 'goto OUT' (which also finalizes the
statement) when they fail.

Note that binding a query parameter failing indicates a programming
error, not a runtime error. Thus, this is a pretty academic fix.

Signed-off-by: Jonny Lamb <jonny.lamb@collabora.co.uk>