Heartbeat: safely stop polling iphb FD.
wocky_heartbeat_source_degrade() is also used to clean up in
wocky_heartbeat_source_finalize(). By the time the latter is called,
WockyHeartbeatSource (which is-a GSource) has been destroyed in the
GSource sense, which implicitly removes all poll FDs. So we need to
guard our call to g_source_remove_poll with a check for whether we've
been destroyed or not.

We could alternatively have two different cleanup functions, one for
when we're degrading and one for when we're being destroyed, but I think
this is simpler.