Porter: fix matching handlers from JIDs sans node part
Previously, registering a handler for a JID with no node part was
accidentally equivalent to registering a handler with from=NULL; that
is, we'd erroneously pass stanzas from *any* server to the handler
function even if it explicitly specified a JID which was just a domain,
as opposed to a JID with an '@' sign in it.

This patch corrects this error, and adds a simple regression test.