WockyXmppReader: Don't drop whitespace-only contents
For some reason, Wocky went out of its way to drop the contents of nodes
if they were purely whitespace. I don't see the point of this, and it
causes user-visible changes. If I send a contact a message containing
only spaces, I'd expect them to receive my message; but because the
<body/> as passed up to Gabble has no contents, Gabble doesn't show it
to the user. Arguably Gabble should show an empty message even if the
body is empty, but if we're doing that we shouldn't mangle
whitespace-only bodies. (Besides, how else am I going to do the
whitespace-only OTR handshake?)

Fixes: <https://bugs.freedesktop.org/show_bug.cgi?id=30042>