porter: don't close twice if we close in a send callback
We need to decide before we complete the "send stanza" asynchronous
operation whether we have a close result pending. If we don't, we can
get into a pickle if the callback to send_async calls close_async. The
problem is that close_async doesn't use the same stanza queue as
send_async so needs special-casing.

For example: Previously, we send a stanza using send_async. The simple
async result would be completed with g_simple_async_result_complete()
and the callback to send_async would be called. However, if in the
callback close_async is called, the stream close is sent directly with
wocky_xmpp_connection_send_close_async and priv->close_result is
assigned. Once the callback returns, send_stanza_cb continues BUT at
the end there is a test for whether there is a close result
pending. This is for when you send a whole bunch of stanzas followed
by a close request, so the porter waits until all the stanzas are sent
before sending the stream close, which is cool. However,
priv->close_result is now != NULL so the porter thinks we're in
exactly that case -- that we're waiting until all the stanzas are sent
and will try and send a stream close afterwards. As we've already done
that the xmpp connection will say "hey another send operation pending"
and everything will fall apart.

Signed-off-by: Jonny Lamb <jonny.lamb@collabora.co.uk>