meta-porter: don't leak the stanza if send_iq_finish wasn't called
We get a new ref to the new stanza from the c2s porter's
send_iq_finish. We now set that as the simple async result's gpointer
with free func of g_object_unref. The finish func returns a new ref to
the stanza so that when the simple async result is completed it unrefs
the stanza leaving the only ref in the callback (unless it has already
been unreffed there).

Bingo.

Signed-off-by: Jonny Lamb <jonny.lamb@collabora.co.uk>