meta-porter: don't assert if we have another connection
It turns out this assertion was being hit a little too often and it
was actually somewhat feasible. Let's clear up new connections when
they're not needed. I thought about adding some kind of support for
multiple porters for this case but it got a bit hairy and I feel that
if you already have an open connection to a contact and another
appears, this is possibly a bug elsewhere.

Signed-off-by: Jonny Lamb <jonny.lamb@collabora.co.uk>