Add an extra_identify argument to the TLS verification functions
We need to be able to pass extra host names that we consider valid to
the code that verifies TLS certificates.
This is useful, for instance, when using Google apps. Rather than
providing a certificate for the JID’s domain part, the server provides a
certificate for talk.google.com; if the user has explicitly configured a
‘Google Talk’ account, it's reasonable to accept certificates for this
domain.