Replace stub DEBUG macros with no-op static inline functions
Previously, if --disable-debug was passed to configure, DEBUG (and
DEBUG_STANZA and DEBUG_NODE_TREE) were defined to expand to
'do { } while (0)'. This meant that the arguments passed to them were
not actually present in the pre-processed C source, which often
triggered warnings about assigning to variables and never reading from
them.

Instead, we define them to be variadic static inline functions. The
compiler should optimize this away, but since we do not use
-Wunused-arguments the call should still count as reading from the
variable, and thus set-but-unused warnings will not be triggered. This
means we don't have to mess around wrapping code to initialize variables
only used in debug messages with #ifdef ENABLE_DEBUG any more. It's true
that, if the compiler is insufficiently smart, this might cause a small
amount of unnecessary work to be done at runtime—for instance,
ping_iq_cb retrieves the from='' attribute of the stanza it's replying
to—but none of these are likely to be too expensive (compared to the
cost of constantly calling wocky_debug(), which is the real reason that
--disable-debug exists).

I am proud of this trick.