Connector tests: always check return when expecting failure
Previously, when we were expecting a specific error code to be returned,
we would not check that the connector had returned NULL rather than a
WockyXmppConnection.

This had the side-effect that, when these cases failed, the test would
segfault (because 'error' was NULL) rather than producing a meaningful
assertion failure.