Add an incomplete write mode to the test-stream
To stress the various bit of code a bit more, make writes by default
never succeed completely. Only write half the buffer that is given to
us as this causes some more code-paths to be tested.

Unfortunately some tests seem to be too sensitive to this. So switch
them back to the old write mode.