Ensure all data in a WockyNode is validated according to the rules in g_utf8_validate
In some cases valid utf-8, but for non-character codepoints gets send
through xmpp. Other parts of the system can get confused by those (e.g.
D-Bus doesn't allow them in strings). So we need to ensure that they get
filtered out as soon as possible.

There are two obvious places to do this, either in the XmppReader or
the Nodes. Doing it in the nodes has the advantage of ensuring
consistency for all data going into the Node, not just what was read
from xmpp directly. Furthermore the Node api calls already copy the
strings internally, which is a good place place to do the conversions if
needed.