dataforms: fix crash in add_field_to_node_using_default()
field->raw_value_contents was not initialized when the field was not coming
from a WockyNode. It leads to a segfault in add_field_to_node_using_default().

The following code was triggering the crash:

  form = g_object_new (WOCKY_TYPE_DATA_FORM, NULL);
  wocky_data_form_set_string (form, "field", "value", TRUE);
  wocky_data_form_add_to_node (form, ...);

This patch initializes field->raw_value_contents when the WockyNode is created
through data_form_set_value().

The unit tests are updated to check this.

https://bugs.freedesktop.org/show_bug.cgi?id=43584