Fix warnings from wocky-connector-test.c under mingw32
In windows socket API setsockopt takes a const char * as the fourth
parameter, while on Linux it takes a void * so we cast to
const char * which will work under both the platforms. This patch
also move flags veriable in client_connected to the unix part of the
ifdef, to remove warnings which shows up with mingw32.