porter-test: fix /handler-stanza
test_close_porter() assumes test->sched_in has not been started, and
calls wocky_xmpp_connection_recv_stanza_async() on test->in; the
callback assumes that the call has failed with an error indicating that
the stream has been closed. This would be fine if the test was sure
there were no stanzas in the queue in test->in, but since my change to
make WockyC2SPorter send back an error in response to unhandled IQs,
this assumption is no longer true: there are error replies sitting
around waiting to be handled.

So this patch just makes this test turn both porters on and then use
test_close_both_porters().