GNUTLS backend: don't use 0 as an error domain
It's never been considered valid, and since GLib 2.31.something
it issues a warning, which breaks the Gabble regression tests.

We're still using 0 as the error code because WockyTLSError has
apparently never actually defined what its error enum is, and in its
existing uses in the OpenSSL backend it's an OpenSSL error code,
which is obviously not applicable here.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=46379
Reviewed-by: Vivek Dasmohapatra <vivek@collabora.co.uk>