Simplify generating GEnums/GFlags
This version of the rule is taken from Gabble. It puts them all together
in one file, and determines the naming from the names of the enumeration
types, rather than from the filenames, which is clearer and saner.

(Under the previous rule, the Jingle code I am about to import ended up
with enums like JINGLE_TYPE_JINGLE_REASON because the file was called
jingle-foo.h and the type was called WockyJingleReason...)