Fix reference to wocky_porter_register_handler()
This was renamed to wocky_porter_register_handler_from(), but its
documentation was not updated. Also, the arguments to the function were
wrong: there was a NULL where the JID should be, and the JID appeared
later in the wrong place.

While we're here, fix some formatting earlier on in the same comment. I
think using &commat; rather than the <!-- --> trick is more legible.