WockyConnector: reduce duplicated code to complete result
Notice that some of the paths set priv->result to NULL before completing
it, and some do not. I particularly like the one in
establish_session_recv_cb which assigns priv->result to a tmp variable
but forgets to set priv->result to NULL!

I hoped this would make the connector re-usable between successful
connection attempts, but it doesn't because priv->conn is not cleared
out so it crashes when the stream is establed the second time you call
wocky_connector_connect_async(). At least all the code paths crash
consistently now, though! (Previously if you had connected to a pre-XMPP
Jabber server, wocky_connector_connect_async() did not immediately
reject a second call and would crash later in the same way this now
does.)