Acknowledge Jingle IQs in a way that the Google webmail client likes
The Google webmail client currently starts calls like this:

    <iq type="set" ...>
        <jingle xmlns="urn:xmpp:jingle:1'"
          action="session-initiate" ...>...</jingle>
        <session xmlns="http://www.google.com/session"
          type="initiate" ...>...</session>
    </iq>

(This isn't valid XMPP Core, because 'An IQ stanza of type "get"
or "set" MUST contain exactly one child element', but we can
tolerate it.)

When called, it also echoes the contents of the sender's IQ back to the
sender. It appears that this is how, when it makes an outgoing call,
it determines which dialect the recipient wants to use: the recipient
echoes the appropriate child element.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=65131
Reviewed-by: Xavier Claessens <xavier.claessens@collabora.co.uk>