Tests: don't assume that other things can sneak in ahead of cancellation
The porter test asserted that if you cancelled the sending of a stanza
after it had already been (sent and) received, the send reported
success, not cancellation; and the SASL auth test asserted that if
you closed a connection at around the same time that a cancellable
had been cancelled, the close reported success, not cancellation.

However, recent GLib seems to be either more careful about deferring the
results of async operations to an idle, or more consistent about
reporting the cancellation as an error even if the operation's
success had already been recorded. As a result, these operations
reported cancellation. To avoid that, delay the cancellation a little.

[Mike Ruprecht clarified on IRC that "more consistent about
reporting the cancellation as an error" is one of the changes from GIO
being ported to GTask in 2.36. -smcv]

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=67900
Signed-off-by: Simon McVittie <simon.mcvittie@collabora.co.uk>
Reviewed-by: Vivek Dasmohapatra <vivek@collabora.co.uk>