Remove wocky-http-proxy, upstreamed in GIO now
The wocky http proxy has now moved in GIO, so we no longer need
to maintain it here. I have kept the unit test for the moment,
to verify that the code in gio works the same as the one that
we used to ship in wocky.

There was one difference in the code from GIO, though. It includes
"Basic" as part of the authorization string. According to RFC this
is correct, so it looks like the wocky proxy client code was doing
this wrong. I have updated the test to reflect that.

Additionally, this commit removes support for building GIO proxy
support depending on whether GIO is recent enough or not.
We can now safely depend unconditionally on a recent enough version.

https://bugs.freedesktop.org/show_bug.cgi?id=94031
Reviewed-by: Diane Trout <diane@ghic.org>