wocky-connector-test: allow test server to dispatch any pending events before teardown
This is to fix an assertion failure that happens in the
"/connector+ssl/econnreset/client-open" test. In this test, the server closes
the connection and the client catches up immediately and closes too.
When the client closes, it forcibly terminates the server, but in most
cases the server's socket close operation doesn't have a chance to finish
before server teardown is called, so the teardown function hits cancel
on the GCancellable that is linked with the operation and finally
the operation ends up repording an "Operation cancelled" error.

By running the GMainLoop once before calling teardown, we ensure
that the socket close operation has a chance to finish gracefully.